keywords

+++ database/migrations/2023_12_12_114315_create_service_user_table.php
@@ -13,17 +13,19 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('service_user', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('service_id');
-            $table->boolean('on_patient_site')->default(0)->nullable();
-            $table->boolean('on_provider_site')->default(0)->nullable();
+        if (!Schema::hasTable('service_user')) {
+            Schema::create('service_user', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('user_id');
+                $table->unsignedBigInteger('service_id');
+                $table->boolean('on_patient_site')->default(0)->nullable();
+                $table->boolean('on_provider_site')->default(0)->nullable();
 
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
-            $table->foreign('service_id')->references('id')->on('services')->onDelete('set null');
-            $table->timestamps();
-        });
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
+                $table->foreign('service_id')->references('id')->on('services')->onDelete('set null');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
